refactor: move transforms to separate file
